requirements.txt
# Minimal requirements for standalone MCP client
# Python 3.11+ required

# MCP Protocol SDK
mcp>=0.9.0

# HTTP client for API requests
httpx>=0.27.0

# Environment variables management
python-dotenv>=1.0.0

